Rewrite cantilevered beam tests
